pyproject.toml
[project]
name = "blur"
version = "0.1.0"
description = ""
authors = [
    {name = "Zhou", email = "lilaboc.cn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "pywin32>=306",
    "pyperclip>=1.8.2",
    "loguru>=0.7.2",
    "moviepy>=1.0.3",
    "scikit-image>=0.24.0",
]

[project.scripts]
blur = "blur.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
